Rename toolbar container class
